+++ src/api/versions.rs
@@ -324,14 +324,14 @@ pub struct VersionWrapper<T> {
 mod test {
     use super::*;
     use crate::api::test_helpers::with_project;
-    use parking_lot::{const_rwlock, RwLock};
     use pretty_assertions::assert_eq;
     use std::error::Error;
+    use tokio::sync::RwLock;
     use tracing_test::traced_test;
 
     /// needed so we do not get 404s when listing while
     /// creating/deleting or creating/updating/deleting
-    static VERSION_LOCK: RwLock<()> = const_rwlock(());
+    static VERSION_LOCK: RwLock<()> = RwLock::const_new(());
 
     #[traced_test]
     #[test]