+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.3.0
+
+add async support (not very well tested yet)
+replace parking_lot lock in tests  with tokio one since the parking_lot one
+does not work with the async tests
+
 ## 0.2.9
 
 update dependencies